--- MyWifiDirect/app/src/main/java/com/example/mywifidirect/MainActivity.java
@@ -1,97 +1,0 @@
-package com.example.mywifidirect;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-
-import android.Manifest;
-import android.app.ActivityManager;
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
-
-public class MainActivity extends AppCompatActivity {
-    private static final String TAG = "MainActivity";
-
-    private static final int REQ_WIFI_DIRECT_PERMISSION = 5566;
-    private static final int HOST_PAGE = 1;
-    private static final int CLIENT_PAGE = 2;
-
-    private static boolean permissionComplete = false;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        Button hostButton = findViewById(R.id.host);
-        hostButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(MainActivity.this, HostActivity.class);
-                startActivityForResult(intent, HOST_PAGE);
-            }
-        });
-
-        Button clientButton = findViewById(R.id.client);
-        clientButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(MainActivity.this, ClientActivity.class);
-                startActivityForResult(intent, HOST_PAGE);
-            }
-        });
-
-        Log.e(TAG, "permission granting");
-        performPermissionGrant();
-    }
-
-    public void performPermissionGrant() {
-        ActivityCompat.requestPermissions(MainActivity.this,
-                new String[]{
-                        Manifest.permission.CHANGE_NETWORK_STATE,
-                        Manifest.permission.ACCESS_NETWORK_STATE,
-                        Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                        Manifest.permission.READ_EXTERNAL_STORAGE,
-                        Manifest.permission.ACCESS_WIFI_STATE,
-                        Manifest.permission.CHANGE_WIFI_STATE,
-                        Manifest.permission.ACCESS_FINE_LOCATION
-                }, REQ_WIFI_DIRECT_PERMISSION);
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (requestCode == REQ_WIFI_DIRECT_PERMISSION) {
-            for (int i = 0; i < grantResults.length; i++) {
-                if (grantResults[i] != PackageManager.PERMISSION_GRANTED) {
-                    showToast("permission failure: " + permissions[i]);
-                    permissionComplete = false;
-                    Log.e(TAG, "set to false");
-                    return;
-                }
-            }
-
-            if (grantResults.length == 0)
-                return;
-
-            showToast("permission pass");
-            //permissionComplete = true;
-            Log.e(TAG, "set to true " + grantResults.length);
-        }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-    }
-
-    private void showToast(String message) {
-        Toast.makeText(this, message, Toast.LENGTH_SHORT).show();
-    }
-}
