--- MyWifiDirect/app/src/main/java/com/example/wifi/SocketConfig.java
@@ -1,6 +1,0 @@
-package com.example.wifi;
-
-public class SocketConfig {
-    public static final int SOCKET_PORT = 9999;
-    public static final int END_CHAR = '$';
-}
