--- MyWifiDirect/app/src/main/java/com/example/wifi/SocketListener.java
@@ -1,9 +1,0 @@
-package com.example.wifi;
-
-import java.net.Socket;
-
-public interface SocketListener {
-    void onAdded(SocketThread socketThread);
-    void onRemoved(SocketThread socketThread);
-    void onRead(SocketThread socketThread, byte[] message);
-}
