--- MyWifiDirect/app/src/main/res/layout/client_main.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <Button
-        android:id="@+id/client_exit"
-        android:layout_width="107dp"
-        android:layout_height="wrap_content"
-        android:text="exit" />
-
-    <Button
-        android:id="@+id/client_start"
-        android:layout_width="112dp"
-        android:layout_height="wrap_content"
-        android:text="Start" />
-
-    <EditText
-        android:id="@+id/client_edit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="edit"
-        android:imeOptions="actionDone"/>
-
-    <Button
-        android:id="@+id/client_to"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="To" />
-
-    <TextView
-        android:id="@+id/client_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="GGGGG" />
-
-</LinearLayout>
