--- MyWifiDirect/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Oct 04 22:11:40 CST 2021
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
